Re-initialize plugin and completers on settings change
Previously changing settings required restart of Sublime which was
rather annoying when testing. Now changes are applied instantly.

Made settings and completer be members of the EasyClangComplete class
as that seems a bit cleaner to me (although one global is still
necessary) and also had to add settings listener which makes sense to
handle in the same class.

Changes Sublime listeners to not be @staticmethods as they need to
access class members settings and completer now. There is IMO no
point in them being static methods anyway as class is only
instantiated once.