Do not mangle order of flags specified through "clang_flags" (#122) (#135)
Use list instead of set for processing includes as set can change the order of
flags an in some cases that can cause compilation errors. For example in case
like that:

'-Xclang', 'find-bad-constructs',

where '-Xclang' needs to be before 'find-bad-constructs' argument.

This change might also affect the order of flags read from generated cmake
compiler_commands file but I would say that this should neither make it worse
nor better. Currently flags for all compilation units are combined into one
set which means that there is not really a proper way of handling order in that
case.