Travis/Appveyor install GNUstep to get ObjC headers
I got some errors on both Linux and Windows in ObjC system headers, i.e.
inside <Foundation/Foundation.h>, but this was tough to debug and
probably not a common use case, so I just worked around in the tests
for now.

Removed "--std=c++11" from .clang_complete because that causes failures
on C and Objective C files. Replaced with -Wabi, which Clang documents
as having no effect; -Wabi still allows us to test for the inclusion
of the flag.