Macro tooltips (#260)
* Refactor ClangUtils::build_info_details to take a cindex module

Instead of a layer of redirection when lib_complete.py builds lists of
types from cindex, just keep a reference to the loaded cindex module
and pass that around. There should be no visible changes from this
commit, but this will make future work on macros easier because there
will more direct access to cindex's types and methods.

* Enable macro tooltips

Hovering over a macro instantiation will show a the macro's signature and
a hyperlink to it's definition.
Examples:
    "#define MACRO 1" will show a tooltip "#define MACRO"
    "#define MACRO() 1" will show a tooltip "#define MACRO()"
    "#define MACRO(x, y) (x + y)" will show a tooltip "#define MACRO(x, y)"

To do this, we first need to parse with PARSE_DETAILED_PROCESSING_RECORD so
libclang will give us macro cursors.

Then, we need to open the file the macro is contained in and parse the
macro definition to get its signature.

Macro doxygen comments will not be shown, but could be added to MacroParser
in the future.