Allow to specify custom header-to-source lookups (#362)
* Allow to specify custom header-to-source lookups

When using a compilation DB (i.e. directly or via CMake), one can now
use templates to find a "related" source file for a header file. This is
done by using globbing expressions which may contain additional
placeholders. These expressions describe how to look for a related
source file relative to a given header file. The first match that is
found in the DB is then used as flags source. This reduced the number of
flags passed to clang for auto-completion which in turn might avoid some
compilation issues in more complex project setups.

* Ignore invalid paths when searching related srcs

When looking up related source files for a header, ignore invalid (None)
paths.

* Use fnmatch to find related source files

Switching to use `fnmatch` instead of `glob`. This basically is required
to allow easier testing (without "real" files on disk). This might also
have a slight advantage regarding performance, as there are no
additional accesses to the disk to find related files - instead, we
search in the already loaded compilation DB for a related file.

* Simplify lookup code for related files + logging

* Allow to use paths in header lookup tables

One can now also use directories (relative to a header file) to find
related source files. In addition, some defaults have been changed so we
maintain the old behavior.